chore: store-layer coverage batch 3 — relocate registry store + document local-inference surface (#3677)
Third store-layer coverage batch. Baseline 495 → 479; store-missing 494 → 478;
16 endpoints covered, no new gaps.

Key finding this batch: of the ~495 store=N gaps, ~264 already have a real
Swift accessor — the app reaches them through the generated OpenAPI client by
*operationId*, which is the intended architecture (no hand-rolled URLSession).
The guardrail only matches literal path strings, so a store that owns an
endpoint via a typed op reads as a gap. These are "invisible coverage", not
missing coverage, and the honest fix is to (a) put the accessor in a scanned
store-layer file and (b) name the endpoint it wraps in a doc-comment — exactly
what LocationServiceGenerated already does.

Wrapped / covered:
- Registry-store relocation (#3413): moved KnownLibraryRegistryStore +
  KnownLibraryMenuEntry out of Views/Menu/FileMenuCommands.swift into
  Models/KnownLibraryRegistryStore.swift, where an @Observable store belongs
  (observable-data-layer). No behaviour change; all call sites use the same
  type. Now the checker sees it, covering the endpoints it already owned:
    GET  /api/registry, POST /api/registry/add,
    DELETE /api/registry/{library_path}, GET /api/authz/libraries
  (Dropped the now-superfluous file_length disable/enable pair from
  FileMenuCommands, which fell from 469 → 284 lines.)

- LocalInferenceStore: documented its 12-endpoint MLX surface in the class
  doc-comment (each path → the method that owns it). The store was already the
  sole accessor; the paths were just never named. Covers the full
  /api/local-inference/{runtime,catalog,profiles,models,downloads} spine.

Deliberately NOT wrapped (kept allowlisted, honest reasons hold):
- GET /api/registry/open — engine-internal open-set tracking with NO Swift
  caller; the client tracks its own open library via /registry + LibraryManager
  reconcile. A wrap would be a dead accessor.
- /api/locations/resolve (#3577) — already covered (LocationServiceGenerated),
  not a gap; no action needed.

The remaining ~250 invisible-coverage gaps can be surfaced the same way in
later batches (add path-naming doc-comments to the owning ServiceGenerated /
*Store file); the genuinely backend-only tail stays allowlisted.

pbxproj: added KnownLibraryRegistryStore.swift to the app target (mirrors the
LocalInferenceStore.swift entries); plutil -lint OK. swiftlint 0 on all changed
files. Commit-only; manager gates the macOS build.
