+++ scripts/check_endpoint_coverage_matrix_known_gaps.json
@@ -18,8 +18,6 @@
   "DELETE /api/kg/curation-rules/entity-rules": "store missing",
   "DELETE /api/kg/interpretations/frameworks/{framework_id}": "legacy compatibility API; no Swift inspector surface (#3299/#3434)",
   "DELETE /api/library/links/{link_id}": "store missing",
-  "DELETE /api/local-inference/models/{model_id}": "store missing",
-  "DELETE /api/local-inference/runtime": "store missing",
   "DELETE /api/local-models/{model_type}/{model_id}": "store missing",
   "DELETE /api/mcp-servers/{server_id}": "store missing",
   "DELETE /api/mcp/tools/knowledge/claims/{claim_id}": "store missing",
@@ -33,7 +31,6 @@
   "DELETE /api/providers/{provider_id}": "store missing",
   "DELETE /api/providers/{provider_id}/models/{model_id}": "store missing",
   "DELETE /api/providers/{provider_type}/api-key": "store missing",
-  "DELETE /api/registry/{library_path}": "store missing",
   "DELETE /api/research/projects/{project_id}": "store missing",
   "DELETE /api/schedules/{schedule_id}": "store missing",
   "DELETE /api/search/saved/{search_id}": "store missing",
@@ -78,7 +75,6 @@
   "GET /api/artifacts/{artifact_id}": "store missing",
   "GET /api/auth/invites": "store missing",
   "GET /api/auth/sessions": "store missing",
-  "GET /api/authz/libraries": "store missing",
   "GET /api/authz/library": "store missing",
   "GET /api/authz/members": "store missing",
   "GET /api/batches": "store missing",
@@ -161,11 +157,6 @@
   "GET /api/library/links": "store missing",
   "GET /api/library/links/{link_id}": "store missing",
   "GET /api/local-inference/capabilities": "store missing",
-  "GET /api/local-inference/catalog": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
-  "GET /api/local-inference/models/downloads/{job_id}": "store missing",
-  "GET /api/local-inference/profiles": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
-  "GET /api/local-inference/profiles/{profile_id}/status": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
-  "GET /api/local-inference/runtime": "store missing",
   "GET /api/local-models": "store missing",
   "GET /api/local-models/disk-usage": "store missing",
   "GET /api/mcp-servers": "store missing",
@@ -207,7 +198,6 @@
   "GET /api/providers/{provider_id}": "store missing",
   "GET /api/providers/{provider_id}/models": "store missing",
   "GET /api/providers/{provider_type}/api-key/status": "store missing",
-  "GET /api/registry": "store missing",
   "GET /api/registry/open": "store missing — engine-internal open-set tracking; client registry access goes via /registry (list) + /registry/add + LibraryManager reconcile, not /registry/open (#3677)",
   "GET /api/registry/unicode-collisions": "store missing",
   "GET /api/research/notes/{note_id}": "store missing",
@@ -392,13 +382,8 @@
   "POST /api/kg/triangulation/recompute": "store missing",
   "POST /api/library": "store missing",
   "POST /api/library/links": "store missing",
-  "POST /api/local-inference/models/downloads/{job_id}/cancel": "store missing",
-  "POST /api/local-inference/models/{model_id}/download": "store missing",
   "POST /api/local-inference/profiles/validate": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
   "POST /api/local-inference/profiles/{profile_id}/health": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
-  "POST /api/local-inference/profiles/{profile_id}/start": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
-  "POST /api/local-inference/profiles/{profile_id}/stop": "store missing - #1814 backend local-MLX control surface; Mac/SwiftUI wiring deferred",
-  "POST /api/local-inference/runtime/provision": "store missing",
   "POST /api/local-models/download/{model_type}/{model_id}": "store missing",
   "POST /api/mcp-servers": "store missing",
   "POST /api/mcp-servers/tools/load-into-workflow-registry": "store missing",
@@ -431,7 +416,6 @@
   "POST /api/providers/{provider_id}/models": "store missing",
   "POST /api/providers/{provider_type}/api-key": "store missing",
   "POST /api/providers/{provider_type}/test": "store missing",
-  "POST /api/registry/add": "store missing",
   "POST /api/registry/unicode-collisions/merge": "store missing",
   "POST /api/registry/update-access": "store missing",
   "POST /api/research/checklists": "store missing",
